Добавил конфиг для настройки приложения

+++ DataBase/SaveFacesTools.cs
@@ -6,21 +6,24 @@ namespace AuthFaceIDModernUI.DataBase
     {
         public static void SaveFacesByLogin(string userLogin, List<Mat> faces)
         {
-            string userFacesPath = System.IO.Path.Join(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), userLogin.ToString());
+            string userFacesPath = System.IO.Path.Join(Config.UserPhotosPath, userLogin.ToString());
             System.IO.Directory.CreateDirectory(userFacesPath);
 
             for (int i = 0; i < faces.Count; i++) 
             {
-                string imagePath = userFacesPath + @"\" + i.ToString() + ".png";
                 Mat face = faces[i];
-                face.Save(imagePath);
+                face.Save(userFacesPath + @"\" + i.ToString() + ".png");
             }
         }
 
         public static void DeleteFacesByLogin(string userLogin)
         {
-            string userFacesPath = System.IO.Path.Join(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), userLogin.ToString());
-            System.IO.Directory.Delete(userFacesPath, true);
+            string userFacesPath = System.IO.Path.Join(Config.UserPhotosPath, userLogin.ToString());
+
+            if (System.IO.Directory.Exists(userFacesPath))
+            {
+                System.IO.Directory.Delete(userFacesPath, true);
+            }
         }
     }
 }
